fix(mining): surface ffmpeg tail error in failure summary (BUG-834)
制卡句子音频失败时 toast/日志只显示 ffmpeg version+configuration banner，
真因（末尾的 Conversion failed! / Encoder not found 等）被从头截断吃掉。
把 extractFfmpegFailureReason 搬到 ffmpeg_backend.dart 并让共享的
FfmpegRunResult.failureSummary 改用它从尾段抽真因（同视频 TODO-910 已修，
但从未惠及音频/reader 制卡路径）。
