+++ hibiki/lib/src/media/video/video_clip_exporter.dart
@@ -3,6 +3,13 @@ import 'dart:io';
 import 'package:hibiki/src/media/video/ffmpeg_backend.dart';
 import 'package:hibiki/src/utils/misc/error_log_service.dart';
 
+// BUG-834：`extractFfmpegFailureReason` 的正准实现搬到 ffmpeg_backend.dart（最底层、
+// 无依赖），供共享的 [FfmpegRunResult.failureSummary] 与音频/视频两路径统一复用。此处
+// 再导出，保持既有 `import '.../video_clip_exporter.dart' show extractFfmpegFailureReason`
+// 调用点与测试不变。
+export 'package:hibiki/src/media/video/ffmpeg_backend.dart'
+    show extractFfmpegFailureReason;
+
 enum VideoClipExportFailure {
   invalidRange,
   inputMissing,
@@ -175,78 +182,6 @@ Future<VideoClipExportResult> exportVideoClipViaFfmpeg({
       detail: e.toString(),
     );
   }
-}
-
-/// 纯函数：从 ffmpeg 的 stderr 里抽取「真正的失败原因行」，给失败 OSD 显示。
-///
-/// 根因（TODO-910）：ffmpeg 的 stderr 开头**恒是输入 banner**——
-/// `Input #0, matroska,webm, from '<path>': Metadata: encoder :...`
-/// （`-hide_banner` 只去版本 banner，不去 `-i` 的输入/流信息）。真正的失败行
-/// （`Conversion failed!` / `Stream map '0:a:3' matches no streams` /
-/// `Could not open file` / `Invalid data found` 等）出现在 stderr **末尾**。
-/// 旧 OSD 用 `substring(0, 160)` 从头截断 → 用户只看到没用的输入 banner，
-/// 看不到真因。本函数从尾段抽真因，让失败提示可读。
-///
-/// 策略（从尾往头扫，跳过 banner/进度/Metadata 噪声行）：
-/// 1. 把 stderr 拆成非空行（去掉行内首尾空白）。
-/// 2. **优先**：自尾向首找第一条「含错误关键词」的行
-///    （error / failed / invalid / could not / no such / matches no streams /
-///     unable / permission denied / not found），返回它。
-/// 3. **退化**：没有任何错误关键词行（如只有 banner），返回**最后一条非噪声行**
-///    （跳过 `Input #` / `Metadata:` / `Stream #` / `Duration:` / `encoder` /
-///     纯进度 `frame=...` 这类信息行）；若全是噪声，返回最后一条非空行。
-/// 4. 全空 → 空串（调用方据此不追加 detail）。
-String extractFfmpegFailureReason(String stderr) {
-  final List<String> lines = stderr
-      .split('\n')
-      .map((String l) => l.trim())
-      .where((String l) => l.isNotEmpty)
-      .toList();
-  if (lines.isEmpty) return '';
-
-  const List<String> errorMarkers = <String>[
-    'error',
-    'failed',
-    'invalid',
-    'could not',
-    'cannot',
-    'no such',
-    'matches no streams',
-    'unable',
-    'permission denied',
-    'not found',
-    'unsupported',
-    'unrecognized',
-    'unknown',
-    'does not contain',
-  ];
-  for (int i = lines.length - 1; i >= 0; i--) {
-    final String lower = lines[i].toLowerCase();
-    if (errorMarkers.any(lower.contains)) {
-      return lines[i];
-    }
-  }
-
-  // 退化：无错误关键词（典型是被截断的纯 banner）。返回最后一条非噪声信息行。
-  bool isNoise(String line) {
-    final String lower = line.toLowerCase();
-    return line.startsWith('Input #') ||
-        line.startsWith('Output #') ||
-        line.startsWith('Stream #') ||
-        line.startsWith('Metadata:') ||
-        line.startsWith('Duration:') ||
-        lower.startsWith('frame=') ||
-        lower.startsWith('size=') ||
-        lower.contains('encoder') ||
-        lower.startsWith('built with') ||
-        lower.startsWith('configuration:') ||
-        lower.startsWith('lib');
-  }
-
-  for (int i = lines.length - 1; i >= 0; i--) {
-    if (!isNoise(lines[i])) return lines[i];
-  }
-  return lines.last;
 }
 
 void _deleteIfPresent(File file) {
